+++ src/App.tsx
@@ -2,8 +2,8 @@ import { Tab, Tabs } from "@nextui-org/tabs";
 import { ReferralTab } from "./tabs/ReferralsTab";
 import { RewardsTab } from "./tabs/RewardsTab";
 import { WidgetContainer } from "./components/TotalNumberWidget/WidgetContainer";
-import { PointsLeaderboard } from "./components/PointsLeaderboard";
-import { ClaimHistory } from "./components/ClaimHistory";
+import { PointsLeaderboard } from "./tabs/PointsLeaderboard";
+import { ClaimHistory } from "./tabs/ClaimHistory";
 
 function App() {
   return (
