+++ src/tabs/PointsLeaderboard/index.ts
